ref(ui): Split ProjectDebugFiles into multiple smaller files
