+++ src/sentry/static/sentry/app/views/settings/projectDebugFiles/types.tsx
@@ -1,0 +1,34 @@
+export enum DebugFileType {
+  EXE = 'exe',
+  DBG = 'dbg',
+  LIB = 'lib',
+}
+
+export enum DebugFileFeature {
+  SYMTAB = 'symtab',
+  DEBUG = 'debug',
+  UNWIND = 'unwind',
+  SOURCES = 'sources',
+}
+
+export type BuiltinSymbolSource = {
+  hidden: boolean;
+  id: string;
+  name: string;
+  sentry_key: string;
+};
+
+export type DebugFile = {
+  codeId: string;
+  cpuName: string;
+  dateCreated: string;
+  debugId: string;
+  headers: Record<string, string>;
+  id: string;
+  objectName: string;
+  sha1: string;
+  size: number;
+  symbolType: string;
+  uuid: string;
+  data?: {type: DebugFileType; features: DebugFileFeature[]};
+};
